+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 # What's new:
-- Added compatibility with EntityCulling's occlusion culling, this means improved performances when both mods are used
-- Improved frustum culling checks by caching bounding boxes
+- Fixed blocks showing up as very big